Split ApiController into Home/Validations/Submissions/Monitoring/Packages
The Phase 1 commit translated all 17 Sinatra routes into one
ApiController as a mechanical step. Split them along resource lines
the way Rails wants them and move the helpers that each controller
actually uses into the controller that owns them. ApplicationController
now carries only cross-cutting helpers (CORS, config accessors,
render_error).

- HomeController: static-file endpoints (/api, /api/apispec, client,
  error_*.json).
- ValidationsController: create / show / status / file / autocorrect,
  plus the upload plumbing that only matters here.
- SubmissionsController: ids / show, with authenticate_curator as a
  before_action.
- MonitoringController: the self-probe biosample round-trip and the
  Net::HTTP helpers it needs.
- PackagesController: list / list_with_groups / attributes /
  attribute_template / info, sharing a Package builder and a package
  param presence check.

Routes remain unchanged; only the controller destination changed.
