+++ config/routes.rb
@@ -2,29 +2,28 @@ Rails.application.routes.draw do
   get '/up' => 'rails/health#show', as: :rails_health_check
 
   scope '/api' do
-    get  ''                                       => 'api#index'
-    get  'apispec/'                               => 'api#apispec'
-    get  'client/index'                           => 'api#client'
+    get  ''                                       => 'home#index'
+    get  'apispec/'                               => 'home#apispec'
+    get  'client/index'                           => 'home#client'
+    get  'error_unauthorized.json'                => 'home#error_unauthorized'
+    get  'error_forbidden.json'                   => 'home#error_forbidden'
+    get  'error_not_found.json'                   => 'home#error_not_found'
 
-    post 'validation'                             => 'api#validation_create'
-    get  'validation/:uuid'                       => 'api#validation_show',        uuid: /[0-9a-f-]+/
-    get  'validation/:uuid/status'                => 'api#validation_status',      uuid: /[0-9a-f-]+/
-    get  'validation/:uuid/:filetype'             => 'api#validation_file',        uuid: /[0-9a-f-]+/
-    get  'validation/:uuid/:filetype/autocorrect' => 'api#validation_autocorrect', uuid: /[0-9a-f-]+/
+    post 'validation'                             => 'validations#create'
+    get  'validation/:uuid'                       => 'validations#show',     uuid: /[0-9a-f-]+/
+    get  'validation/:uuid/status'                => 'validations#status',   uuid: /[0-9a-f-]+/
+    get  'validation/:uuid/:filetype'             => 'validations#file',     uuid: /[0-9a-f-]+/
+    get  'validation/:uuid/:filetype/autocorrect' => 'validations#autocorrect', uuid: /[0-9a-f-]+/
 
-    get  'submission/ids/:filetype'               => 'api#submission_ids'
-    get  'submission/:filetype/:submission_id'    => 'api#submission_show',        submission_id: /[^\/]+/
+    get  'submission/ids/:filetype'               => 'submissions#ids'
+    get  'submission/:filetype/:submission_id'    => 'submissions#show',     submission_id: /[^\/]+/
 
-    get  'monitoring'                             => 'api#monitoring'
+    get  'monitoring'                             => 'monitoring#show'
 
-    get  'package_list'                           => 'api#package_list'
-    get  'package_and_group_list'                 => 'api#package_and_group_list'
-    get  'attribute_list'                         => 'api#attribute_list'
-    get  'attribute_template_file'                => 'api#attribute_template_file'
-    get  'package_info'                           => 'api#package_info'
-
-    get  'error_unauthorized.json'                => 'api#error_unauthorized'
-    get  'error_forbidden.json'                   => 'api#error_forbidden'
-    get  'error_not_found.json'                   => 'api#error_not_found'
+    get  'package_list'                           => 'packages#list'
+    get  'package_and_group_list'                 => 'packages#list_with_groups'
+    get  'attribute_list'                         => 'packages#attributes'
+    get  'attribute_template_file'                => 'packages#attribute_template'
+    get  'package_info'                           => 'packages#info'
   end
 end
